:fire: chore(viz): retire built-in p5 sketch classes — DEFAULT_VIZ_DESCRIPTORS compiles from bundled source (#184)
Problem: the viz picker mounted 7 hand-written TypeScript sketch classes
(PianorollSketch, ScopeSketch, FscopeSketch, SpectrumSketch, SpiralSketch,
PitchwheelSketch, WordfallSketch) while the workspace preset files
(`preset/viz/*.p5`) carried separate source strings. After #178 fixed the
runtime resolution so `.viz("pianoroll")` reaches the preset (PV56 at
resolution time), the sketch classes were dead code — but they still
shipped, and the duality "what the picker mounts ≠ what the user edits"
could regrow at any time.

Fix: PV56 fully satisfied. One physical source of truth per viz.

  • New `packages/editor/src/visualizers/builtinP5Code.ts` carries the
    7 p5 source strings (PIANOROLL_P5_CODE, SCOPE_P5_CODE,
    FSCOPE_P5_CODE, SPECTRUM_P5_CODE, SPIRAL_P5_CODE, PITCHWHEEL_P5_CODE,
    WORDFALL_P5_CODE). Verbatim move from `packages/app/src/templates.ts`.
  • `DEFAULT_VIZ_DESCRIPTORS` p5 entries now build their factories via
    `new P5VizRenderer(compileP5Code(CODE, 'id'))` instead of
    `new P5VizRenderer(SketchClass)`. Same `P5SketchFactory` shape on
    both sides; the picker and the preset file mount the same code.
  • `packages/editor/src/index.ts` re-exports the 7 constants and drops
    the 5 sketch class re-exports.
  • `packages/app/src/templates.ts` and
    `packages/app/src/components/StrudelEditorClient.tsx` import the
    constants from `@stave/editor` instead of carrying local copies.
  • Deletes the 7 sketch class files (~890 LOC).
  • Deletes `PianorollSketch.test.ts` (the pure utils it tested are
    inlined into the bundled p5 code where they cannot be unit-tested
    without spinning up p5; behavior is exercised by the existing
    `viz-single-source.spec.ts` visual gate).
  • Drops the dead `PianorollSketch` mock in `viewZones.test.ts`.

Note: `StrudelEditorClient.seedPresets` still overwrites the persisted
Piano Roll preset code with `PIANOROLL_P5_CODE` on every load (clobbering
user edits to the bundled viz). That predates this PR and is filed as
its own issue per the #184 body — out of scope here.

Validation:

  • 1662 editor unit tests green (-21 deleted PianorollSketch helper
    tests, no new tests needed — `defaultDescriptors.test.ts` already
    walks every entry's `factory()` which now exercises the compile
    path; `vizCompiler.test.ts` covers `compileP5Code` directly).
  • `pnpm --filter @stave/app build` clean (P72 — full tsc, BUILD_ID
    written).
  • `viz-single-source.spec.ts` 2/2 green in Playwright: inline
    `.viz("pianoroll")` mounts a canvas with > 1 distinct color, page
    has ZERO errors, AND the backdrop `.pianoroll()` likewise. The
    >1-distinct-color assertion guards against P74 (consolidation
    downgrade) — if the compiled code had silently lost the
    fold-lane/classification/active-highlight logic, the test would
    fail.
  • P66 grep gate: `PIANOROLL_P5_CODE` + `compileP5Code(PIANOROLL`
    present in `dist/index.js`; sketch class names absent (0 hits).

Closes #184.
