+++ packages/editor/src/visualizers/builtinP5Code.ts
@@ -1,0 +1,370 @@
+/**
+ * Bundled p5 viz source code — single source of truth.
+ *
+ * These strings power both:
+ *   - `DEFAULT_VIZ_DESCRIPTORS` (the viz picker / demo / standalone
+ *     embedders), compiled via `compileP5Code`.
+ *   - The bundled `preset/viz/*.p5` workspace files seeded by the app
+ *     template, which re-export these constants for backwards-compat.
+ *
+ * Until #184 (PR retiring the 7 TS sketch classes that previously powered
+ * the picker), the picker rendered a different — richer or poorer — copy
+ * of each viz than what users edited in the preset file. PV56 demands the
+ * picker code path === the preset file code path; this module is that path.
+ */
+
+export const PIANOROLL_P5_CODE = `// Stave p5 viz — Piano Roll
+// stave.scheduler, stave.analyser, stave.hapStream are injected globals.
+// Fold-by-pitch lanes: each distinct pitch (or unpitched sound) gets its own
+// horizontal lane, sorted low→high, so notes never overlap and the melodic
+// contour reads as a staircase. Notes scroll right→left across a 4-cycle
+// window; the playhead sits at the half mark.
+
+const CYCLES = 4      // cycles visible across the canvas width
+const PLAYHEAD = 0.5  // 0..1 — where "now" sits horizontally
+// Drum/percussion sound-name prefixes for color classification.
+const DRUM_PREFIXES = ['bd', 'sd', 'hh', 'rim', 'cp', 'cy', 'lt', 'mt', 'ht', 'oh', 'cl']
+
+function isDrum(s) {
+  return DRUM_PREFIXES.some(p => s === p || (s.startsWith(p) && /\\d/.test(s[p.length] || '')))
+}
+
+// Note NAME → MIDI. Returns null for unparseable names (octaveless or
+// sample names) — the caller folds those onto string lanes instead.
+function noteToMidi(n) {
+  if (typeof n === 'number') return Math.round(n)
+  if (typeof n !== 'string') return null
+  const m = n.toLowerCase().match(/^([a-g])(b|#)?(-?\\d+)$/)
+  if (!m) return null
+  const base = { c: 0, d: 2, e: 4, f: 5, g: 7, a: 9, b: 11 }[m[1]]
+  const acc = m[2] === 'b' ? -1 : m[2] === '#' ? 1 : 0
+  return (parseInt(m[3]) + 1) * 12 + base + acc
+}
+
+// Fold-grouping key: a MIDI number for pitched haps, a "_sound" string for
+// unpitched ones. Priority mirrors how Strudel fills a hap (freq is
+// pre-computed from note, so note("c e g") resolves via freq, never NaN).
+function valueOf(h) {
+  if (typeof h.freq === 'number') return Math.round(12 * Math.log2(h.freq / 440) + 69)
+  if (typeof h.note === 'number') return h.note
+  if (typeof h.note === 'string') {
+    const mi = noteToMidi(h.note)
+    return mi !== null ? mi : '_' + h.note
+  }
+  if (h.s) return '_' + h.s
+  return 0
+}
+
+function parseHex(hex) {
+  const s = String(hex).replace('#', '')
+  if (s.length === 6) return [parseInt(s.slice(0, 2), 16), parseInt(s.slice(2, 4), 16), parseInt(s.slice(4, 6), 16)]
+  if (s.length === 3) return [parseInt(s[0] + s[0], 16), parseInt(s[1] + s[1], 16), parseInt(s[2] + s[2], 16)]
+  return null
+}
+
+// Explicit hap color wins; else classify by sound family.
+function colorOf(h) {
+  if (h.color) { const c = parseHex(h.color); if (c) return c }
+  const s = h.s || ''
+  if (isDrum(s)) return [249, 115, 22]        // drums  — orange
+  if (s.startsWith('bass')) return [6, 182, 212]   // bass  — cyan
+  if (s.startsWith('pad')) return [16, 185, 129]   // pad   — green
+  return [139, 92, 246]                        // melody — purple
+}
+
+function setup() {
+  createCanvas(stave.width, stave.height)
+  noStroke()
+}
+
+function draw() {
+  const W = width, H = height
+  background(9, 9, 18)
+  const sched = stave.scheduler
+  if (!sched) return
+  let now
+  try { now = sched.now() } catch (e) { return }
+
+  const from = now - CYCLES * PLAYHEAD
+  const to = now + CYCLES * (1 - PLAYHEAD)
+  const ext = to - from
+  let haps
+  try { haps = sched.query(from, to) } catch (e) { haps = [] }
+
+  // Fold: collect distinct pitch/sound values, sort ascending → lanes.
+  const seen = new Set(), vals = []
+  for (const h of haps) { const v = valueOf(h); if (!seen.has(v)) { seen.add(v); vals.push(v) } }
+  vals.sort((a, b) => {
+    if (typeof a === 'number' && typeof b === 'number') return a - b
+    if (typeof a === 'number') return -1
+    if (typeof b === 'number') return 1
+    return String(a).localeCompare(String(b))
+  })
+  const fold = Math.max(1, vals.length)
+  const barH = H / fold
+
+  noStroke()
+  for (const h of haps) {
+    const lane = vals.indexOf(valueOf(h))
+    if (lane < 0) continue
+    const x = ((h.begin - now + CYCLES * PLAYHEAD) / ext) * W
+    const w = Math.max(2, ((h.end - h.begin) / ext) * W)
+    const y = ((fold - 1 - lane) / fold) * H   // higher pitch → higher up
+    const endC = h.endClipped != null ? h.endClipped : h.end
+    const active = h.begin <= now && endC > now
+    const gain = Math.min(1, Math.max(0.1, h.gain == null ? 1 : h.gain))
+    const vel = Math.min(1, Math.max(0.1, h.velocity == null ? 1 : h.velocity))
+    const alpha = gain * vel
+    const [r, g, b] = colorOf(h)
+    if (active) {
+      // Brightened toward white + a crisp outline so the playing note pops.
+      fill(min(255, r + 60), min(255, g + 60), min(255, b + 60), alpha * 255)
+      rect(x, y + 1, w - 2, barH - 2)
+      noFill(); stroke(255, 255, 255, 220); strokeWeight(1)
+      rect(x, y + 1, w - 2, barH - 2); noStroke()
+    } else {
+      fill(r, g, b, alpha * 180)
+      rect(x, y + 1, w - 2, barH - 2)
+    }
+  }
+
+  // Playhead line.
+  stroke(255, 255, 255, 128); strokeWeight(1)
+  line(PLAYHEAD * W, 0, PLAYHEAD * W, H); noStroke()
+}`;
+
+export const SCOPE_P5_CODE = `// Stave p5 viz — Scope (oscilloscope / event pulses)
+function setup() {
+  createCanvas(stave.width, stave.height)
+  noFill()
+}
+function draw() {
+  background(9, 9, 18)
+  stroke(40, 50, 70); strokeWeight(0.5)
+  line(0, height * 0.75, width, height * 0.75)
+  if (stave.analyser) {
+    const buf = stave.analyser.frequencyBinCount
+    const data = new Float32Array(buf)
+    stave.analyser.getFloatTimeDomainData(data)
+    let trig = 0
+    for (let i = 1; i < buf; i++) { if (data[i-1] > 0 && data[i] <= 0) { trig = i; break } }
+    stroke('#75baff'); strokeWeight(2); beginShape()
+    for (let i = trig; i < buf; i++) vertex((i - trig) * width / (buf - trig), (0.75 - 0.25 * data[i]) * height)
+    endShape()
+  } else if (stave.scheduler) {
+    const now = stave.scheduler.now()
+    const haps = stave.scheduler.query(now - 4, now + 0.1)
+    noStroke()
+    for (const h of haps) {
+      const age = now - h.begin, decay = max(0, 1 - age / 4)
+      const x = ((h.begin - now + 4) / 4) * width
+      const w = max(3, ((h.end - h.begin) / 4) * width)
+      const pH = height * 0.6 * decay * (h.gain ?? 1)
+      fill(117, 186, 255, decay * 200)
+      rect(x, height * 0.75 - pH / 2, w, pH, 2)
+    }
+  }
+}`;
+
+export const FSCOPE_P5_CODE = `// Stave p5 viz — Frequency Scope (FFT bars / note bars)
+function setup() {
+  createCanvas(stave.width, stave.height)
+  noStroke()
+}
+function draw() {
+  background(9, 9, 18)
+  stroke(40, 50, 70); strokeWeight(0.5); noFill()
+  line(0, height * 0.75, width, height * 0.75); noStroke()
+  if (stave.analyser) {
+    const buf = stave.analyser.frequencyBinCount
+    const data = new Float32Array(buf)
+    stave.analyser.getFloatFrequencyData(data)
+    fill('#75baff')
+    const sw = width / buf
+    for (let i = 0; i < buf; i++) {
+      const n = constrain((data[i] + 100) / 100, 0, 1), v = n * 0.25
+      rect(i * sw, (0.75 - v * 0.5) * height, max(sw, 1), v * height)
+    }
+  } else if (stave.scheduler) {
+    const now = stave.scheduler.now()
+    const haps = stave.scheduler.query(now - 0.2, now + 0.05)
+    const bins = new Float32Array(64)
+    for (const h of haps) {
+      const note = typeof h.note === 'string' ? 60 : (h.note ?? 60)
+      const freq = 440 * pow(2, (note - 69) / 12)
+      if (freq < 30) continue
+      const idx = constrain(floor(log(freq / 30) / log(4000 / 30) * 64), 0, 63)
+      bins[idx] = max(bins[idx], max(0, 1 - (now - h.begin) / 0.5) * (h.gain ?? 1))
+    }
+    const sw = width / 64
+    for (let i = 0; i < 64; i++) {
+      if (bins[i] <= 0) continue
+      const v = bins[i] * 0.25
+      fill(117, 186, 255, bins[i] * 220)
+      rect(i * sw, (0.75 - v * 0.5) * height, max(sw - 1, 1), v * height)
+    }
+  }
+}`;
+
+export const SPECTRUM_P5_CODE = `// Stave p5 viz — Spectrum (scrolling waterfall)
+function setup() {
+  createCanvas(stave.width, stave.height)
+  pixelDensity(1); noStroke()
+}
+function draw() {
+  const ctx = drawingContext
+  if (stave.analyser) {
+    const buf = stave.analyser.frequencyBinCount
+    const data = new Float32Array(buf)
+    stave.analyser.getFloatFrequencyData(data)
+    const img = ctx.getImageData(0, 0, width, height)
+    ctx.clearRect(0, 0, width, height)
+    ctx.putImageData(img, -2, 0)
+    ctx.fillStyle = '#75baff'
+    for (let i = 0; i < buf; i++) {
+      const n = constrain((data[i] + 80) / 80, 0, 1)
+      if (n <= 0) continue
+      ctx.globalAlpha = n
+      const yEnd = (log(i + 1) / log(buf)) * height
+      const yStart = i > 0 ? (log(i) / log(buf)) * height : 0
+      ctx.fillRect(width - 2, height - yEnd, 2, max(2, yEnd - yStart))
+    }
+    ctx.globalAlpha = 1
+  } else if (stave.scheduler) {
+    const now = stave.scheduler.now()
+    const img = ctx.getImageData(0, 0, width, height)
+    ctx.clearRect(0, 0, width, height)
+    ctx.putImageData(img, -2, 0)
+    const haps = stave.scheduler.query(now - 0.3, now + 0.05)
+    for (const h of haps) {
+      const note = typeof h.note === 'string' ? 60 : (h.note ?? 60)
+      const freq = 440 * pow(2, (note - 69) / 12)
+      if (freq < 20) continue
+      const logPos = log(freq / 20) / log(4000 / 20)
+      const y = height - logPos * height
+      const alpha = max(0.1, 1 - (now - h.begin) / 0.5) * (h.gain ?? 1)
+      ctx.fillStyle = h.color ?? '#75baff'
+      ctx.globalAlpha = alpha
+      ctx.fillRect(width - 2, y - 2, 2, max(4, height * 0.03))
+    }
+    ctx.globalAlpha = 1
+  } else { background(9, 9, 18) }
+}`;
+
+export const SPIRAL_P5_CODE = `// Stave p5 viz — Spiral
+function setup() {
+  createCanvas(300, 200)
+  pixelDensity(window.devicePixelRatio || 1)
+  noFill()
+}
+function xySpiral(rot, margin, cx, cy, rotate) {
+  const a = ((rot + rotate) * 360 - 90) * PI / 180
+  return [cx + cos(a) * margin * rot, cy + sin(a) * margin * rot]
+}
+function draw() {
+  background(9, 9, 18)
+  if (!stave.scheduler) return
+  const now = stave.scheduler.now()
+  const haps = stave.scheduler.query(now - 2, now + 1)
+  const cx = width / 2, cy = height / 2
+  const sz = min(width, height) * 0.38, mg = sz / 3
+  for (const h of haps) {
+    const active = h.begin <= now && h.end > now
+    const from = h.begin - now + 3, to = h.end - now + 3 - 0.005
+    const op = max(0, 1 - abs((h.begin - now) / 2))
+    const c = color(h.color ?? (active ? '#75baff' : '#8a919966'))
+    c.setAlpha(op * 255)
+    stroke(c); strokeWeight(mg / 2); strokeCap(ROUND)
+    beginShape()
+    for (let a = from; a <= to; a += 1/60) {
+      const [x, y] = xySpiral(a, mg, cx, cy, now)
+      vertex(x, y)
+    }
+    endShape()
+  }
+  stroke(255); strokeWeight(mg / 2)
+  beginShape()
+  for (let a = 2.98; a <= 3; a += 1/60) {
+    const [x, y] = xySpiral(a, mg, cx, cy, now)
+    vertex(x, y)
+  }
+  endShape()
+}`;
+
+export const PITCHWHEEL_P5_CODE = `// Stave p5 viz — Pitchwheel
+const ROOT_FREQ = 440 * pow(2, (36 - 69) / 12)
+function setup() {
+  createCanvas(300, 200)
+  pixelDensity(window.devicePixelRatio || 1)
+}
+function freq2angle(f) { return 0.5 - (log(f / ROOT_FREQ) / log(2) % 1) }
+function circPos(cx, cy, r, a) {
+  const rad = a * TWO_PI
+  return [sin(rad) * r + cx, cos(rad) * r + cy]
+}
+function draw() {
+  background(9, 9, 18)
+  if (!stave.scheduler) return
+  const now = stave.scheduler.now()
+  let haps = stave.scheduler.query(now - 0.01, now + 0.01)
+  haps = haps.filter(h => h.begin <= now && h.end > now)
+  const sz = min(width, height), r = sz / 2 - 12
+  const cx = width / 2, cy = height / 2
+  noStroke(); fill(117, 186, 255, 64)
+  for (let i = 0; i < 12; i++) {
+    const a = freq2angle(ROOT_FREQ * pow(2, i / 12))
+    const [x, y] = circPos(cx, cy, r, a)
+    circle(x, y, 7)
+  }
+  noFill(); stroke(117, 186, 255, 48); strokeWeight(1)
+  circle(cx, cy, r * 2)
+  for (const h of haps) {
+    const note = typeof h.note === 'string' ? 60 : (h.note ?? 60)
+    const freq = 440 * pow(2, (note - 69) / 12)
+    const a = freq2angle(freq)
+    const [x, y] = circPos(cx, cy, r, a)
+    const c = h.color ?? '#75baff'
+    stroke(c); strokeWeight(2)
+    line(cx, cy, x, y)
+    fill(c); noStroke()
+    circle(x, y, 12)
+  }
+}`;
+
+export const WORDFALL_P5_CODE = `// Stave p5 viz — Wordfall (vertical pianoroll with labels)
+function setup() {
+  createCanvas(stave.width, stave.height)
+  pixelDensity(window.devicePixelRatio || 1)
+}
+function draw() {
+  background(9, 9, 18)
+  if (!stave.scheduler) return
+  const now = stave.scheduler.now()
+  const CYCLES = 4, PH = 0.5
+  const haps = stave.scheduler.query(now - CYCLES * PH, now + CYCLES * (1 - PH))
+  const vals = [...new Set(haps.map(h => h.note ?? h.s ?? 0))].sort()
+  if (!vals.length) return
+  const bw = width / vals.length
+  for (const h of haps) {
+    const active = h.begin <= now && h.end > now
+    const dur = h.end - h.begin
+    const yOff = h.begin - now
+    const y = height * PH - (yOff / CYCLES) * height
+    const dH = (dur / CYCLES) * height
+    const v = h.note ?? h.s ?? 0
+    const x = vals.indexOf(v) * bw
+    noStroke()
+    if (active) fill(255)
+    else { const c = color(h.color ?? '#75baff'); c.setAlpha(160); fill(c) }
+    rect(x + 1, y + 1, bw - 2, dH - 2)
+    if (dH > 10 && bw > 16) {
+      const label = h.note != null ? String(h.note) : (h.s ?? '')
+      textSize(min(bw * 0.55, dH * 0.7, 11))
+      textAlign(LEFT, TOP); fill(active ? 0 : 255); noStroke()
+      text(label, x + 3, y + 3)
+    }
+  }
+  stroke(255, 255, 255, 128); strokeWeight(1)
+  line(0, height * PH, width, height * PH)
+}`;
+
